+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/dao/HibernateBaseDao.java
@@ -21,6 +21,7 @@ import java.util.Objects;
  * @author DwArFeng
  * @since 0.0.1-beta
  */
+@SuppressWarnings("DuplicatedCode")
 public class HibernateBaseDao<K extends Key, PK extends Bean, E extends Entity<K>, PE extends Bean> implements BaseDao<K, E> {
 
     private HibernateTemplate template;